Add files via upload
webserver.py is a simple apache2 webserver I created using the python language on Google Cloud's VM service. The webserver uses a TCP connection and listens on port 12000. The web client will send a GET request to the webserver for a specific html file, and if the file is accessable to the webserver, it will return it to client. If no such file exists, the webserver will send html displaying to the client that no file exists.