Add files via upload
proxy.py is a simple apache2 proxy server I created using the python language on Google Cloud's VM service. The proxy uses a TCP connection and listens on port 12001. A web client will send a request for a specific website, and the proxy will receive the GET request from the client. The proxy server will use that GET request to retrieve the desired website and return it to the user.