Create folder for listener.py and sender.py
sender.py and listener.py are an example of a simple UDP connection, where the listner.py file will listen on all IP addresses and port 12000 for any incoming traffic. The sender.py file will send out a message on port 12000, which the listener.py file will pick up and display the information of the message to the screen. The listener.py will send an acknowledgment back to the sender.py file on what it received.